# docs/swagger.yaml
basePath: /api
definitions:
  router.CreateDictEntityRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
  router.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      teamId:
        type: integer
      username:
        type: string
    type: object
  router.DictEntityResponse:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  router.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  router.ListDictEntitiesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/router.DictEntityResponse'
        type: array
    type: object
  router.ListUsersResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/router.UserResponse'
        type: array
    type: object
  router.UpdateDictEntityRequest:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  router.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      roles:
        items:
          $ref: '#/definitions/router.DictEntityResponse'
        type: array
      teamId:
        type: integer
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Incident management system API
  title: Fixik API
  version: "1.0"
paths:
  /impacts:
    get:
      consumes:
      - application/json
      description: Get all impacts in dictionary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ListDictEntitiesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List impacts
      tags:
      - impacts
    post:
      consumes:
      - application/json
      description: Create new impact dictionary entry
      parameters:
      - description: Impact data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create impact
      tags:
      - impacts
...
  /priorities:
    get:
      consumes:
      - application/json
      description: Get all priorities in dictionary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ListDictEntitiesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List priorities
      tags:
      - priorities
    post:
      consumes:
      - application/json
      description: Create new status dictionary entry
      parameters:
      - description: Priority data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create status
      tags:
      - priorities
...
  /roles:
    get:
      consumes:
      - application/json
      description: Get all roles in dictionary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ListDictEntitiesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create new role dictionary entry
      parameters:
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create role
      tags:
      - roles
...
  /statuses:
    get:
      consumes:
      - application/json
      description: Get all statuses in dictionary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ListDictEntitiesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List statuses
      tags:
      - statuses
    post:
      consumes:
      - application/json
      description: Create new status dictionary entry
      parameters:
      - description: Status data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create status
      tags:
      - statuses
  /statuses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete status dictionary entry by ID
      parameters:
      - description: Status ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Delete status
      tags:
      - statuses
    get:
      consumes:
      - application/json
      description: Get status dictionary entry by ID
      parameters:
      - description: Status ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Get status by ID
      tags:
      - statuses
    put:
      consumes:
      - application/json
      description: Update status dictionary entry by ID
      parameters:
      - description: Status ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.UpdateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Update status
      tags:
      - statuses
  /teams:
    get:
      consumes:
      - application/json
      description: Get all teams in dictionary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ListDictEntitiesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Create new team dictionary entry
      parameters:
      - description: Team data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create team
      tags:
      - teams
  /teams/{id}:
    delete:
      consumes:
      - application/json
      description: Delete team dictionary entry by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Delete team
      tags:
      - teams
    get:
      consumes:
      - application/json
      description: Get team dictionary entry by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Get team by ID
      tags:
      - teams
    put:
      consumes:
      - application/json
      description: Update team dictionary entry by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: Team data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.UpdateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Update team
      tags:
      - teams
  /users:
    get:
      consumes:
      - application/json
      description: Get all users with optional limit and offset
      parameters:
      - default: 100
        description: Limit of items
        in: query
        name: limit
        required: true
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ListUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create new user
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create user
      tags:
      - users
...
schemes:
- http
swagger: "2.0"